backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pypdf==5.0.0
google-genai==0.3.0
python-dotenv==1.0.1
motor==3.6.0
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
